refs #10 : Updated

+++ lib/model/doctrine/PlugincrMediaGalleryContent.class.php
@@ -13,12 +13,26 @@
 abstract class PlugincrMediaGalleryContent extends BasecrMediaGalleryContent
 {
   /**
-   * Resize image after upload
+   * Resize image after upload or delete files id content_delete
+   * is checked
    * @param object $event
    */
   public function postSave($event)
   {
     $this->resizeImage();
+  }
+
+  public function preSave($event)
+  {
+    if(!$this->isNew())
+    {
+      $request = sfContext::getInstance()->getRequest();
+      $form = $request->getParameter('cr_media_gallery_content');
+      if(isset($form['content_delete']))
+      {
+        $this->deleteContent();
+      }
+    }
   }
 
   public function preDelete($event)
@@ -43,11 +57,14 @@ abstract class PlugincrMediaGalleryContent extends BasecrMediaGalleryContent
       $image = $full_dir.DS.$this->getContent();
       $thumb = $full_dir_thumb.DS.$this->getContent();
 
-      $size = $this->getResizeOptions($form['resize_w'],$form['resize_h']);
+      if(file_exists($image)&&!is_dir($image))
+      {
+        $size = $this->getResizeOptions($form['resize_w'],$form['resize_h']);
 
-      $thumbnail = new sfThumbnail($size[0], $size[1]);
-      $thumbnail->loadFile($image);
-      $thumbnail->save($thumb, 'image/jpeg');
+        $thumbnail = new sfThumbnail($size[0], $size[1]);
+        $thumbnail->loadFile($image);
+        $thumbnail->save($thumb, 'image/jpeg');
+      }
     }
   }
 
@@ -82,25 +99,20 @@ abstract class PlugincrMediaGalleryContent extends BasecrMediaGalleryContent
     $full_dir = sfConfig::get('sf_upload_dir').DS.'galleries'.DS.$gaid.DS.$type.DS;
     $full_dir_thumb = $full_dir.DS.'thumbnail'.DS;
 
-    if(file_exists($full_dir.$cont))
+    if(file_exists($full_dir.$cont) && !is_dir($full_dir.$cont))
     {
       if(!unlink($full_dir.$cont))
       {
         throw new sfException('Can\'t delete file: '.$full_dir.$cont);
       }
     }
-    else
-    {
-      die($full_dir.$cont);
-    }
-
-    if('image'==$type&&file_exists($full_dir_thumb.$cont))
+    
+    if('image'==$type && file_exists($full_dir_thumb.$cont) && !is_dir($full_dir_thumb.$cont))
     {
       if(!unlink($full_dir_thumb.$cont))
       {
         throw new sfException('Can\'t delete file: '.$full_dir_thumb.$cont);
       }
     }
-
   }
 }
